frontend/requirements.txt
uvicorn==0.22.0
fastapi==0.95.1
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.24.0
python-dotenv==1.0.0
openai>=1.0.0,<2.0.0
aiofiles==23.1.0